requirements.txt
requests==2.31.0
networkx==3.2.1
python-bitcoinrpc==1.0
python-dotenv==1.0.0
